+++ src/server/controlchan/line_parser/mod.rs
@@ -1,6 +1,6 @@
 //! This modules implements a line parser for FTP control channel commands
 //!
-//! Use the parse method. It takes a FTP line and returns an instance of the Command enum.
+//! Use the parse method. It takes an FTP line and returns an instance of the Command enum.
 //!
 pub mod error;
 mod parser;